+++ src/Pages/Dashboard/AdminDashboard/AdminHome/AdminHome.jsx
@@ -1,0 +1,11 @@
+
+
+const AdminHome = () => {
+    return (
+        <div>
+           <h2 className="text-2xl">Admin home</h2> 
+        </div>
+    );
+};
+
+export default AdminHome;
