--- src/Pages/Dashboard/AdminHome/AdminHome.jsx
@@ -1,11 +1,0 @@
-
-
-const AdminHome = () => {
-    return (
-        <div>
-           <h2 className="text-2xl">Admin home</h2> 
-        </div>
-    );
-};
-
-export default AdminHome;
